separating function logic to utils
