+++ tfbpshiny/utils/rank_response_replicate_plot_utils.py
@@ -1,0 +1,304 @@
+# %%
+import pandas as pd
+import plotly.graph_objects as go
+from scipy.stats import binom, binomtest
+from scipy.stats._result_classes import BinomTestResult
+
+
+def parse_binomtest_results(binomtest_obj: BinomTestResult, **kwargs):
+    """
+    Parses the results of a binomtest into a tuple of floats.
+
+    This function takes the results of a binomtest and returns a tuple of
+    floats containing the response ratio, p-value, and confidence interval
+    bounds.
+
+    :param binomtest_obj: The results of a binomtest for a single rank bin.
+        Additional keyword arguments: Additional keyword arguments are passed
+        to the proportional_ci method of the binomtest object.
+
+    :return: A tuple of floats containing the response ratio, p-value, and
+        confidence interval bounds.
+
+    :examples:
+
+    .. code-block:: python
+
+        from scipy.stats import binomtest
+        from rank_response_plot.logic import parse_binomtest_results
+
+        result = binomtest(k=1, n=2, p=0.5, alternative='greater')
+        parse_binomtest_results(result, confidence_level=0.95)
+        # Output: (0.5, <p-value>, <ci_lower>, <ci_upper>)
+
+    """
+    return (
+        binomtest_obj.statistic,
+        binomtest_obj.pvalue,
+        binomtest_obj.proportion_ci(
+            confidence_level=kwargs.get("confidence_level", 0.95),
+            method=kwargs.get("method", "exact"),
+        ).low,
+        binomtest_obj.proportion_ci(
+            confidence_level=kwargs.get("confidence_level", 0.95),
+            method=kwargs.get("method", "exact"),
+        ).high,
+    )
+
+
+def compute_rank_response(df: pd.DataFrame, **kwargs) -> pd.DataFrame:
+    """
+    Computes rank-based statistics and binomial test results for a DataFrame.
+
+    This function groups the DataFrame by 'rank_bin' and aggregates it to
+    calculate the number of responsive items in each rank bin, as well as
+    various statistics related to a binomial test.  It calculates the
+    cumulative number of successes, response ratio, p-value, and confidence
+    intervals for each rank bin.
+
+    :param df: DataFrame containing the columns 'rank_bin',
+        'responsive', and 'random'. 'rank_bin' is an integer representing
+        the rank bin, 'responsive' is a boolean indicating responsiveness,
+        and 'random' is a float representing the random expectation.
+    :param kwargs: Additional keyword arguments are passed
+        to the binomtest function, including arguments to the
+        proportional_ci method of the BinomTestResults object (see scipy
+        documentation for details)
+
+    :return: A DataFrame indexed by 'rank_bin' with columns for the
+            number of responsive items in each bin ('n_responsive_in_rank'),
+            cumulative number of successes ('n_successes'), response ratio
+            ('response_ratio'), p-value ('p_value'), and confidence interval
+            bounds ('ci_lower' and 'ci_upper').
+
+    :examples:
+
+    .. code-block:: python
+        df = pd.DataFrame({'rank_bin': [1, 1, 2],
+        ...                    'responsive': [True, False, True],
+        ...                    'random': [0.5, 0.5, 0.5]})
+        compute_rank_response(df)
+        # Returns a DataFrame with rank-based statistics and binomial
+        # test results.
+
+    """
+    rank_response_df = (
+        df.groupby("rank_bin")
+        .agg(
+            n_responsive_in_rank=pd.NamedAgg(column="responsive", aggfunc="sum"),
+            random=pd.NamedAgg(column="random", aggfunc="first"),
+        )
+        .reset_index()
+    )
+
+    rank_response_df["n_successes"] = rank_response_df["n_responsive_in_rank"].cumsum()
+
+    # Binomial Test and Confidence Interval
+    rank_response_df[["response_ratio", "pvalue", "ci_lower", "ci_upper"]] = (
+        rank_response_df.apply(
+            lambda row: parse_binomtest_results(
+                binomtest(
+                    int(row["n_successes"]),
+                    int(row.rank_bin),
+                    float(row["random"]),
+                    alternative=kwargs.get("alternative", "two-sided"),
+                ),
+                **kwargs,
+            ),
+            axis=1,
+            result_type="expand",
+        )
+    )
+
+    return rank_response_df
+
+
+def binom_ci(trials, random_prob, alpha=0.05):
+    """
+    Calculate the confidence interval for a binomial distribution. This function
+    calculates the confidence interval for a binomial distribution using the binomial
+    cumulative distribution function (CDF).
+
+    :param trials: The number of trials (n).
+    :param random_prob: The probability of success (p).
+    :param alpha: The significance level (default is 0.05).
+    :return: A tuple containing the lower and upper bounds of the confidence interval.
+
+    """
+    lower_bound = binom.ppf(alpha / 2, trials, random_prob) / trials
+    upper_bound = binom.ppf(1 - alpha / 2, trials, random_prob) / trials
+    return lower_bound, upper_bound
+
+
+def process_plot_data(data: pd.DataFrame, key: str) -> dict:
+    """
+    Process the data for plotting. This function filters the data for a specific key,
+    computes rank-based statistics, and prepares the data for plotting.
+
+    :param key: The key for the data to be processed.
+    :param data: The DataFrame containing the data to be processed.
+    :return: A dictionary containing the processed data for plotting.
+
+    """
+    subset_data = data[data["rank_bin"] <= 150]
+    rr_summary = compute_rank_response(subset_data)
+    return {
+        "x": rr_summary["rank_bin"],
+        "y": rr_summary["response_ratio"],
+        "random_y": rr_summary.get("random"),
+        "ci": (
+            rr_summary["rank_bin"].apply(
+                lambda n: binom_ci(n, rr_summary["random"].iloc[0])
+            )
+            if "random" in rr_summary
+            else None
+        ),
+    }
+
+
+def prepare_rank_response_data(rr_dict: dict) -> dict:
+    """
+    Prepare rank response data for plotting.
+
+    :param rr_dict: Dictionary containing rank response data.
+    :return: Dictionary containing processed data for plotting.
+
+    """
+    metadata: pd.DataFrame = rr_dict.get("metadata")
+    data_dict: dict = rr_dict.get("data", {})
+    plots: dict = {}
+    for _, row in metadata.iterrows():
+        expression_id = str(row["expression"])
+        promotersetsig_id = str(row["promotersetsig"])
+        data = data_dict.get(str(row["id"]))
+        if data is not None:
+            plots.setdefault(expression_id, {})[promotersetsig_id] = process_plot_data(
+                promotersetsig_id, data
+            )
+    return plots
+
+
+def add_traces_to_plot(fig, sample_id, add_random, **kwargs):
+    """Add traces to a Plotly figure based on plot data."""
+    # Add the main line for the promoterset signal
+    fig.add_trace(
+        go.Scatter(
+            x=kwargs["x"],
+            y=kwargs["y"],
+            mode="lines",
+            name=f"{sample_id}",
+            legendrank=int(sample_id),
+        )
+    )
+
+    if add_random:
+        # Add the random line
+        fig.add_trace(
+            go.Scatter(
+                x=kwargs["x"],
+                y=kwargs["random_y"],
+                mode="lines",
+                name="Random",
+                line=dict(dash="dash", color="black"),
+                legendrank=0,
+            )
+        )
+
+        if kwargs["ci"] is not None:
+            ci_lower = kwargs["ci"].apply(lambda x: x[0])
+            ci_upper = kwargs["ci"].apply(lambda x: x[1])
+
+            # Add confidence interval lower bound
+            fig.add_trace(
+                go.Scatter(
+                    x=kwargs["x"],
+                    y=ci_lower,
+                    mode="lines",
+                    line=dict(width=0),
+                    showlegend=False,
+                )
+            )
+
+            # Add confidence interval upper bound and shade the area
+            fig.add_trace(
+                go.Scatter(
+                    x=kwargs["x"],
+                    y=ci_upper,
+                    mode="lines",
+                    fill="tonexty",
+                    fillcolor="rgba(128, 128, 128, 0.3)",
+                    line=dict(width=0),
+                    showlegend=False,
+                )
+            )
+
+
+def create_rank_response_replicate_plot(plots_dict):
+    """Generate a dictionary of Plotly figures from the prepared rank response data."""
+    output_dict = {}
+
+    for expression_id, promotersetsig_dict in plots_dict.items():
+        fig = go.Figure()
+        add_random = True
+        for promotersetsig_id, plot_data in promotersetsig_dict.items():
+            # Use the helper function to add traces to the plot
+            add_traces_to_plot(fig, promotersetsig_id, add_random, **plot_data)
+            add_random = False  # Add random line only once
+
+        # Update the layout of the figure
+        fig.update_layout(
+            title={
+                "text": f"Rank Response for Expression ID {expression_id}",
+                "x": 0.5,
+            },
+            yaxis_title="# Responsive / # Genes",
+            xaxis_title="Number of Genes, Ranked by Binding Score",
+            xaxis=dict(tick0=0, dtick=5, range=[0, 150]),  # Set x-axis ticks and range
+            yaxis=dict(
+                tick0=0, dtick=0.1, range=[0, 1.0]
+            ),  # Set y-axis ticks and range
+        )
+
+        output_dict[expression_id] = fig
+
+    return output_dict
+
+
+# ## Example
+
+# %%
+# Set up the environment
+# import dotenv
+
+# from tfbpapi import *
+
+# dotenv.load_dotenv("/home/chase/code/tfbpshiny/.env", override=True)
+
+# # configure the logger to print to console
+# import logging
+
+# logging.basicConfig(level=logging.DEBUG)
+
+# rr_api = RankResponseAPI()
+
+# # %%
+# rr_api.pop_params()
+# rr_api.push_params(
+#     {
+#         "regulator_symbol": "OAF1",
+#         "expression_conditions": "expression_source=mcisaac_oe,time=15"
+#     }
+# )
+
+# rr_dict = await rr_api.read(retrieve_files=True)
+
+# # %%
+# plots = prepare_rank_response_data(rr_dict)
+
+# # %%
+# x = create_rank_response_replicate_plot(plots)
+
+# %%
+# to show data, do x.get(<id>).show()
+
+# %%
